refactor : 3단계, 4단계 InputView,OutPutView 분리 및 main 함수 ResultView로 이동
