+++ src/main/java/racingCar/RacingCar.java
@@ -3,28 +3,17 @@ package racingCar;
 import racingCar.domain.Car;
 import racingCar.domain.Cars;
 import racingCar.domain.RandomNumber;
-import racingCar.view.InputView;
 import racingCar.view.ResultView;
+import racingCar.view.WinnerResultView;
 
 import java.util.List;
 
 public class RacingCar {
     public static final int RANDOM_LIMIT_VALUE = 10;
 
-    public static void main(String[] args) {
-        int[] inputValues = InputView.input();
-        int carNum = inputValues[0];
-        int tryNum = inputValues[1];
-
-        Cars cars = new Cars(carNum);
-        movingResult(cars, tryNum);
-
-    }
-
     public static Cars movingResult(Cars cars, int tryNum){
         for(int i=0; i< tryNum; i++){
             addCarMoving(cars.getCarList());
-            ResultView.printNewLine();
         }
 
         return cars;
@@ -34,8 +23,8 @@ public class RacingCar {
 
         for(Car car : carList){
             car.moveCar(RandomNumber.getRandom(RANDOM_LIMIT_VALUE));
-            ResultView.printCar(car);
         }
+        ResultView.printCar(carList);
 
         return carList;
     }
@@ -44,7 +33,6 @@ public class RacingCar {
     public static Cars movingNamedCarResult(Cars cars, int tryNum){
         for(int i=0; i< tryNum; i++){
             addNamedCarMoving(cars.getCarList());
-            ResultView.printNewLine();
         }
 
         return cars;
@@ -54,10 +42,18 @@ public class RacingCar {
 
         for(Car car : carList){
             car.moveCar(RandomNumber.getRandom(RANDOM_LIMIT_VALUE));
-            ResultView.printNamedCar(car);
         }
+        WinnerResultView.printNamedCar(carList);
 
         return carList;
+    }
+
+    public static List<String> returnWinner(Cars cars){
+        List<String> winners = cars.returnWinnerCars();
+
+        WinnerResultView.printWinner(winners);
+
+        return winners;
     }
 
 }
