--- src/main/java/racingCar/WinnerCar.java
@@ -1,28 +1,0 @@
-package racingCar;
-
-import racingCar.domain.Cars;
-import racingCar.view.InputView;
-import racingCar.view.ResultView;
-
-import java.util.List;
-import java.util.Map;
-
-public class WinnerCar {
-    public static void main(String[] args) {
-        Map<String, Object> inputMap = InputView.inputNamedCar();
-        String carNames = inputMap.get("carNames").toString();
-        int tyrNum = (int) inputMap.get("tryNum");
-
-        Cars cars = new Cars(carNames);
-        RacingCar.movingNamedCarResult(cars, tyrNum);
-        WinnerCar.returnWinner(cars);
-    }
-
-    public static List<String> returnWinner(Cars cars){
-        List<String> winners = cars.returnWinnerCars();
-
-        ResultView.printWinner(winners);
-
-        return winners;
-    }
-}
