+++ src/main/java/racingCar/view/ResultView.java
@@ -1,37 +1,34 @@
 package racingCar.view;
 
+import racingCar.RacingCar;
 import racingCar.domain.Car;
+import racingCar.domain.Cars;
 
 import java.util.List;
 
 public class ResultView {
 
-    private ResultView(){}
+    public static void main(String[] args) {
+        int[] inputValues = InputView.input();
+        int carNum = inputValues[0];
+        int tryNum = inputValues[1];
 
-    public static void printCar(Car car){
+        Cars cars = new Cars(carNum);
+        RacingCar.movingResult(cars, tryNum);
+
+    }
+
+    public static void printEachCar(Car car){
         for(int i=0; i < car.getForwardCount(); i++){
             System.out.print("-");
         }
         System.out.println();
     }
 
-    public static void printNamedCar(Car car){
-        System.out.print(car.getCarName() + " : ");
-        for(int i=0; i < car.getForwardCount(); i++){
-            System.out.print("-");
+    public static void printCar(List<Car> carList){
+        for(Car car : carList) {
+            printEachCar(car);
         }
-        System.out.println();
-    }
-
-    public static void printWinner(List<String> winners){
-        for(int i=0; i < winners.size()-1; i++){
-            System.out.print(winners.get(i) + ", ");
-        }
-        System.out.print(winners.get(winners.size()-1) + "가 최종 우승했습니다.");
-
-    }
-
-    public static void printNewLine(){
         System.out.println();
     }
